+++ CHANGELOG.md
@@ -6,6 +6,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### 🔄 Changed
 
+# [1.22.2](https://github.com/GetStream/stream-video-swift/releases/tag/1.22.2)
+_May 13, 2025_
+
+### 🐞 Fixed
+- Fix an issue that was causing CallSettings misalignment during reconnection [#810](https://github.com/GetStream/stream-video-swift/pull/810)
+- Synchronize CallKit audioSession with the audioSession in the app. [#807](https://github.com/GetStream/stream-video-swift/pull/807)
+
 # [1.22.1](https://github.com/GetStream/stream-video-swift/releases/tag/1.22.1)
 _May 08, 2025_